+++ src/protocol.c
@@ -1259,7 +1259,7 @@ NOEXPORT void ntlm(CLI *c, SERVICE_OPTIONS *opt) {
         else if(is_prefix(line, "Content-Length: ")) {
             content_length=strtol(line+16, &tmpstr, 10);
             if(tmpstr>line+16) /* found some digits */
-                while(*tmpstr && isspace(*tmpstr))
+                while(*tmpstr && isspace((int)*tmpstr))
                     ++tmpstr;
             if(tmpstr==line+16 || *tmpstr || content_length<0) {
                 s_log(LOG_ERR, "Proxy-Authenticate: Invalid Content-Length");
